requirements.txt
Flask==3.0.2
ytmusicapi==1.5.2
yt-dlp==2024.3.10
ffmpeg-python==0.2.0